+++ extensions/gitlab/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Search Scope] - 2023-04-27
 
-- Add scope dropdown for `Search Issues`
+- Add scope dropdown for `Search Issues` and `Search Merge Requests`
 - Add possibility to search for specific `state`, `author`, `assignees`, `milestone` or `label` in `Search Issues`.
 
 ## [Remove Branch] - 2023-04-17